+++ lib/braintree/address_gateway.rb
@@ -76,8 +76,8 @@ module Braintree
 
     def self._shared_signature # :nodoc:
       [:company, :country_code_alpha2, :country_code_alpha3, :country_code_numeric,
-        :country_name, :extended_address, :first_name,
-        :last_name, :locality, :postal_code, :region, :street_address]
+        :country_name, :extended_address, :first_name, :last_name, :locality, :phone_number,
+        :postal_code, :region, :street_address]
     end
 
     def self._update_signature # :nodoc: