+++ lib/braintree/risk_data.rb
@@ -2,10 +2,13 @@ module Braintree
   class RiskData # :nodoc:
     include BaseModule
 
+    attr_reader :customer_device_id
+    attr_reader :customer_location_zip
+    attr_reader :customer_tenure
     attr_reader :decision
     attr_reader :device_data_captured
-    attr_reader :id
     attr_reader :fraud_service_provider
+    attr_reader :id
 
     def initialize(attributes)
       set_instance_variables_from_hash attributes unless attributes.nil?