+++ lib/braintree/transaction/address_details.rb
@@ -3,6 +3,15 @@ module Braintree
     class AddressDetails # :nodoc:
       include BaseModule
 
+      module ShippingMethod
+        SameDay = "same_day"
+        NextDay = "next_day"
+        Priority = "priority"
+        Ground = "ground"
+        Electronic = "electronic"
+        ShipToStore = "ship_to_store"
+      end
+
       attr_reader :company
       attr_reader :country_code_alpha2
       attr_reader :country_code_alpha3
@@ -13,8 +22,10 @@ module Braintree
       attr_reader :id
       attr_reader :last_name
       attr_reader :locality
+      attr_reader :phone_number
       attr_reader :postal_code
       attr_reader :region
+      attr_reader :shipping_method
       attr_reader :street_address
 
       def initialize(attributes)