+++ lib/braintree/transaction_gateway.rb
@@ -190,15 +190,16 @@ module Braintree
         :shipping_amount, :discount_amount, :ships_from_postal_code,
         :billing_address_id, :payment_method_nonce, :three_d_secure_token, :three_d_secure_authentication_id,
         :shared_payment_method_token, :shared_billing_address_id, :shared_customer_id, :shared_shipping_address_id, :shared_payment_method_nonce,
+        :product_sku,
         {:line_items => [:quantity, :name, :description, :kind, :unit_amount, :unit_tax_amount, :total_amount, :discount_amount, :tax_amount, :unit_of_measure, :product_code, :commodity_code, :url]},
-        {:risk_data => [:customer_browser, :customer_ip]},
+        {:risk_data => [:customer_browser, :customer_device_id, :customer_ip, :customer_location_zip, :customer_tenure]},
         {:credit_card => [:token, :cardholder_name, :cvv, :expiration_date, :expiration_month, :expiration_year, :number]},
         {:customer => [:id, :company, :email, :fax, :first_name, :last_name, :phone, :website]},
         {
           :billing => AddressGateway._shared_signature
         },
         {
-          :shipping => AddressGateway._shared_signature
+          :shipping => AddressGateway._shared_signature + [:shipping_method],
         },
         {
           :three_d_secure_pass_thru => [