+++ spec/unit/braintree/dispute_spec.rb
@@ -319,6 +319,7 @@ describe Braintree::Dispute do
 
       dispute.transaction.amount.should == 31.00
       dispute.transaction.id.should == "open_disputed_transaction"
+      dispute.transaction.created_at.should == Time.utc(2009, 2, 9, 12, 59, 59)
       dispute.transaction.order_id.should == nil
       dispute.transaction.purchase_order_number.should == "po"
       dispute.transaction.payment_instrument_subtype.should == "Visa"